Alteração da solução principal para worker background service
