+++ Docker/docker-compose-devenv.override.yml
@@ -1,14 +1,8 @@
 version: '3.4'
 
 services:
-  FIAP.Pos.Hackathon.Micro.Servico.Processamento.Imagens.Producao.api:
+  FIAP.Pos.Hackathon.Micro.Servico.Processamento.Imagens.Producao.Worker:
     environment:
       - ASPNETCORE_ENVIRONMENT=Development
-      - ASPNETCORE_HTTP_PORTS=8082
-      - ASPNETCORE_HTTPS_PORTS=8083
-    ports:
-      - "8082"
-      - "8083"
     volumes:
       - ${APPDATA}/Microsoft/UserSecrets:/home/app/.microsoft/usersecrets:ro
-      - ${APPDATA}/ASP.NET/Https:/home/app/.aspnet/https:ro
