+++ Docker/docker-compose-devenv.yml
@@ -5,14 +5,14 @@ volumes:
   azurite_data: {}
 
 services:
-  FIAP.Pos.Hackathon.Micro.Servico.Processamento.Imagens.Producao.api:
-    image: ${DOCKER_REGISTRY-}fiapposhackathonmicroservicoprocessamentoimagensproducaoapi
+  FIAP.Pos.Hackathon.Micro.Servico.Processamento.Imagens.Producao.Worker:
+    image: ${DOCKER_REGISTRY-}fiapposhackathonmicroservicoprocessamentoimagensproducaoworker
     # depends_on:
       # - mongodb
       # - azurite
     build:
       context: ../
-      dockerfile: Src/Api/Dockerfile
+      dockerfile: Src/Worker/Dockerfile
   # mongodb:
   #   image: mongo:latest
   #   ports:
