+++ Docker/docker-compose.dcproj
@@ -5,8 +5,6 @@
     <DockerTargetOS>Linux</DockerTargetOS>
     <DockerPublishLocally>False</DockerPublishLocally>
     <ProjectGuid>06e7e3e8-59f1-45f0-a9bd-fac3ab7ad0e4</ProjectGuid>
-    <DockerLaunchAction>LaunchBrowser</DockerLaunchAction>
-    <DockerServiceUrl>{Scheme}://localhost:{ServicePort}/api-docs</DockerServiceUrl>
     <DockerServiceName>FIAP.Pos.Hackathon.Micro.Servico.Processamento.Imagens.Producao.api</DockerServiceName>
     <DockerComposeProjectName>fiap-pos-hackathon-micro-servico-processamento-imagens-producao-gurpo-71</DockerComposeProjectName>
     <DockerComposeBaseFilePath>docker-compose-devenv</DockerComposeBaseFilePath>
