+++ FIAP-Pos-Hackathon-Micro-Servico-Processamento-Imagens-Producao.sln
@@ -11,7 +11,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "External", "External", "{39
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{BB1E9158-AE8F-42EE-B69A-17FB18E5327C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Api", "Src\Api\Api.csproj", "{28ACBA4F-879C-4520-8F1D-5300F1292459}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Worker", "Src\Worker\Worker.csproj", "{28ACBA4F-879C-4520-8F1D-5300F1292459}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Infra", "Src\External\Infra\Infra.csproj", "{21491D01-0CB3-4AC4-B2FA-9D192619DE0E}"
 EndProject
