--- Src/Api/Api.http
@@ -1,6 +1,0 @@
-@Api_HostAddress = http://localhost:5014
-
-GET {{Api_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
